Derive UserKey from an abstract class (with the same behaviour)
This brings no functional change (and no migration), but allows a user to
inherit from the abstract class in order to override fields or change
their behaviour.  This is similar to the AbstractUser and User classes in
Django's auth module.

This allows a lot of flexibility, for instance adding extra constraints or
validation, while keeping the amount of code duplication low.